+++ src/validators/isFutureDate.ts
@@ -1,7 +1,11 @@
 import { parseDate } from "../helpers";
 import { DateLike } from "../types";
 
-export const isFutureDate = (arg: DateLike): boolean => {
+export const isFutureDate = (
+  arg: DateLike,
+  { referenceDate }: { referenceDate?: DateLike } = {}
+): boolean => {
   const date = parseDate(arg);
-  return !!date && date > new Date();
+
+  return !!date && date > (parseDate(referenceDate) ?? new Date());
 };