+++ src/validators/isPastDate.ts
@@ -1,7 +1,11 @@
 import { parseDate } from "../helpers";
 import { DateLike } from "../types";
 
-export const isPastDate = (arg: DateLike): boolean => {
+export const isPastDate = (
+  arg: DateLike,
+  { referenceDate }: { referenceDate?: DateLike } = {}
+): boolean => {
   const date = parseDate(arg);
-  return !!date && date < new Date();
+
+  return !!date && date < (parseDate(referenceDate) ?? new Date());
 };